Consolidate files related to the implementation of the Adaptive Random Forest - Moved ARFHoeffdingTreeClassifier and ARFHoeffdingTreeRegressor into adaptive_random_forest.py - Rename ARFHoeffdingTreeClassifier as BaseTreeClassifier - Rename ARFHoeffdingTreeRegressor as BaseTreeRegressor
